move quote logic
